requirements.txt
aiohttp>=3.7.4
appdirs==1.4.4
APScheduler==3.8.1
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
decorator==5.1.1
discord-webhook==0.14.0
discord.py==1.7.3
dogpile.cache==1.1.5
facepy==1.0.12
Flask==2.0.2
Flask-Markdown==0.3
fuzzywuzzy==0.18.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
msgpack==1.0.3
multidict==6.0.2
musicbrainzngs==0.7.1
mypy-extensions==0.4.3
numpy==1.26.4
opencv-python==4.5.5.62
packaging==21.3
pathspec==0.9.0
pathvalidate==2.5.0
pbr==5.8.0
Pillow==9.0.1
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pylast==4.5.0
pyparsing==3.0.7
pysubs2
pytest==7.0.0
python-dotenv==0.19.2
python-Levenshtein==0.12.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
# PyYAML==6.0
requests==2.27.1
ripgrepy==1.1.0
six==1.16.0
srt==3.5.1
stevedore==3.5.0
tabulate==0.8.9
tmdbsimple==2.9.1
tomli==2.0.1
typing_extensions
tzdata==2021.5
tzlocal==4.1
Unidecode==1.3.2
urllib3==1.26.8
waitress==2.1.2
Wand==0.6.7
Werkzeug==2.0.2
yarl
timeago==1.0.15
SQLAlchemy==2.0.25
pyyaml
lxml
yt-dlp
dynaconf
fastapi
uvicorn
httpx
git+https://github.com/vitiko98/libgenmics
lyricsgenius
requests-cache
alembic
plexapi
deluge-client
cleanit
git+https://github.com/vitiko98/fese
av
